chore(tarcap): clang format + updated paths for renamed/moved files
