+++ src/network/tarcap/packets_handlers/common/packet_handler.cpp
@@ -1,6 +1,6 @@
 #include "packet_handler.hpp"
 
-#include "network/tarcap/packets_handlers/common/packets_stats/packets_stats.hpp"
+#include "network/tarcap/stats/packets_stats.hpp"
 
 namespace taraxa::network::tarcap {
