+++ src/network/tarcap/shared_states/peers_state.cpp
@@ -46,9 +46,8 @@ std::vector<dev::p2p::NodeID> PeersState::getAllPendingPeersIDs() const {
   std::vector<dev::p2p::NodeID> peers;
 
   std::shared_lock lock(peers_mutex_);
-  std::transform(
-      pending_peers_.begin(), pending_peers_.end(), std::back_inserter(peers),
-      [](std::pair<const dev::p2p::NodeID, std::shared_ptr<TaraxaPeer>> const &peer) { return peer.first; });
+  std::transform(pending_peers_.begin(), pending_peers_.end(), std::back_inserter(peers),
+                 [](std::pair<const dev::p2p::NodeID, std::shared_ptr<TaraxaPeer>> const& peer) { return peer.first; });
 
   return peers;
 }
