+++ src/network/tarcap/shared_states/peers_state.hpp
@@ -4,7 +4,7 @@
 #include "libp2p/Common.h"
 #include "libp2p/Host.h"
 #include "network/tarcap/packet_types.hpp"
-#include "network/tarcap/packets_handlers/common/packets_stats/packets_stats.hpp"
+#include "network/tarcap/stats/packets_stats.hpp"
 #include "network/tarcap/taraxa_peer.hpp"
 #include "transaction_manager/transaction.hpp"
